First version of main
Several improvements are clearly possible. Changing the arrays into keyed dictionaries will allow for easier sort than the 5 if statements currently available. The best day/worst day metric still needs to be implemented. However, with the flat gain/percentage gain arrays already completed this should be completed in a single method creation. This version compared to the other, function based software is much easier to work with/on.